requirements/prod.txt
cryptography>=35.0.0,<36.0.0
https://github.com/cheese-drawer/lib-python-db-wrapper/releases/download/2.3.0/db_wrapper-2.3.0-py3-none-any.whl
fastapi>=0.70.0,<0.80.0
python-dotenv>=0.19.2,<0.20.0
python-jose>=3.3.0,<3.4.0
python-multipart==0.0.5
uvicorn[standard]>=0.15.0,<0.16.0
